+++ psd-web/app/controllers/concerns/secondary_authentication_concern.rb
@@ -7,12 +7,12 @@
 module SecondaryAuthenticationConcern
   extend ActiveSupport::Concern
 
-  def require_secondary_authentication
+  def require_secondary_authentication(redirect_to: request.fullpath)
     return unless Rails.configuration.secondary_authentication_enabled
 
     if user_id_for_secondary_authentication && !secondary_authentication_present?
       user = User.find(user_id_for_secondary_authentication)
-      session[:secondary_authentication_redirect_to] = request.fullpath
+      session[:secondary_authentication_redirect_to] = redirect_to
       session[:secondary_authentication_user_id] = user_id_for_secondary_authentication
       auth = SecondaryAuthentication.new(user)
       auth.generate_and_send_code(current_operation)