+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2020-05-06
+- Allow to request a new secondary authentication code when secondary authentication is required and code was not received via sms.
+
 ## 2020-04-27
 - Allow to report a case as safe. Previously one could only report a case as non-compliant, unsafe or both.